fix: Modify structure
