+++ app/(monitoring)/d/content/dataBinding.tsx
@@ -1,3 +1,4 @@
+import LargeBtn from "@/components/button/largeBtn";
 import RoundToggleBtnGroup from "@/components/button/toggle/roundToggleBtnGroup";
 import Dropdown from "@/components/dropdown/dropdown";
 import MultiSelectDropdown from "@/components/dropdown/multiSelectDropdown";
@@ -6,8 +7,20 @@ import TextArea from "@/components/textarea/textarea";
 import { useChartOptions } from "@/context/chartOptionContext";
 import { useSelectedSection } from "@/context/selectedSectionContext";
 import { useWidgetOptions } from "@/context/widgetOptionContext";
-import { MOCK_DATA, COLUMNS } from "@/data/dataBinding";
+import {
+  MOCK_DATA,
+  COLUMNS,
+  metricOptionsMap,
+  TARGET_OPTIONS,
+} from "@/data/dataBinding";
 import React, { useState, useEffect, useMemo } from "react";
+
+const BASE_API_URL = "/api/v1/dataBinding";
+
+const makeApiUrl = (target: string, metric: string) => {
+  // return `${BASE_API_URL}?target=${target}&metric=${metric}`;
+  return `${BASE_API_URL}/${target}`;
+};
 
 const DataBinding = () => {
   const { datasets, setDatasets } = useChartOptions();
@@ -19,8 +32,30 @@ const DataBinding = () => {
   const [selectedColumns, setSelectedColumns] = useState<string[]>([]);
   const [selectedColumn, setSelectedColumn] = useState<string | null>(null);
   const [query, setQuery] = useState<string>("");
-  const [apiUrl, setApiUrl] = useState<string>("");
   const [queryParams, setQueryParams] = useState<string>("");
+
+  const [apiUrl, setApiUrl] = useState<string>(BASE_API_URL);
+  const [selectedTarget, setSelectedTarget] = useState<string>("");
+  const [selectedMetric, setSelectedMetric] = useState<string[]>([]);
+
+  // URL 생성
+  const generatedUrl = useMemo(() => {
+    if (!selectedTarget || selectedMetric.length === 0) return "";
+
+    const params = new URLSearchParams({
+      target: selectedTarget,
+      metric: selectedMetric.join(","),
+    });
+
+    return `/api?${params.toString()}`;
+  }, [selectedTarget, selectedMetric]);
+
+  useEffect(() => {
+    if (selectedTarget) {
+      const url = makeApiUrl(selectedTarget, selectedMetric.join(","));
+      setApiUrl(url);
+    }
+  }, [selectedTarget, selectedMetric]);
 
   // 테이블 선택 핸들러 (테이블 변경 시 컬럼 초기화)
   const handleTableChange = (value: string) => {
@@ -135,9 +170,9 @@ const DataBinding = () => {
   return (
     <div
       className="bg-modern-bg border-l border-0.5 no-scrollbar border-modern-border
-    pt-[44px] pl-6 w-[300px] h-[100vh] overflow-y-auto"
+    pl-6 py-8 w-[300px] h-[100vh] overflow-y-auto"
     >
-      <div className="w-full flex flex-col pt-10 pb-10">
+      <div className="w-full flex flex-col">
         <h2 className="text-lg font-semibold mb-4 text-modern-text">
           Data Binding
         </h2>
@@ -159,20 +194,46 @@ const DataBinding = () => {
         {/* API Binding 옵션 */}
         {isApiBinding === "API" ? (
           <>
-            {/* API URL 입력 */}
-            <div className="flex flex-col mb-6">
-              <label className="text-sm2 text-modern-text mb-2">API URL</label>
-              <TextInput
-                value={apiUrl}
-                onChange={(e) => setApiUrl(e)}
-                placeholder="Enter API URL"
+            <div className="flex flex-col mb-6" style={{ marginBottom: 24 }}>
+              <label className="mb-2 text-sm2 text-modern-text">
+                Generated URL
+              </label>
+              <div className="p-2 text-modern-point bg-modern-point_20 rounded text-sm break-all w-[250px]">
+                {apiUrl || "Please select both Target and Metric."}
+              </div>
+            </div>
+
+            {/* Target 선택 */}
+            <div className="flex flex-col mb-6" style={{ marginBottom: 24 }}>
+              <label className="mb-2 text-sm2 text-modern-text">
+                * Select Target
+              </label>
+              <Dropdown
+                value={selectedTarget}
+                onChange={(value) => {
+                  setSelectedTarget(value);
+                  setSelectedMetric([]);
+                }}
+                options={TARGET_OPTIONS}
                 className="w-[250px]"
               />
             </div>
 
-            {/* Query parameters */}
-            <div className="flex flex-col mb-6">
-              <label className="text-sm2 text-modern-text mb-2">
+            {/* Metric 선택 */}
+            <div className="flex flex-col mb-6" style={{ marginBottom: 24 }}>
+              <label className="mb-2 text-sm2 text-modern-text">
+                * Select Metric
+              </label>
+              <MultiSelectDropdown
+                value={selectedMetric}
+                onChange={setSelectedMetric}
+                options={metricOptionsMap[selectedTarget] || []}
+                placeholder="Select one or more columns"
+                className="w-[250px] min-h-8"
+              />
+            </div>
+            <div className="flex flex-col mb-6" style={{ marginBottom: 24 }}>
+              <label className="mb-2 text-sm2 text-modern-text">
                 Query Parameters (Optional)
               </label>
               <TextInput
@@ -182,6 +243,37 @@ const DataBinding = () => {
                 className="w-[250px]"
               />
             </div>
+
+            <LargeBtn
+              title={" Fetch Data"}
+              onClick={() => {
+                console.log("Fetching from URL:", generatedUrl);
+              }}
+              disabled={!selectedTarget || selectedMetric.length === 0}
+            />
+            {/* API URL 입력 */}
+            {/* <div className="flex flex-col mb-6">
+              <label className="text-sm2 text-modern-text mb-2">API URL</label>
+              <TextInput
+                value={apiUrl}
+                onChange={(e) => setApiUrl(e)}
+                placeholder="Enter API URL"
+                className="w-[250px]"
+              />
+            </div> */}
+
+            {/* Query parameters */}
+            {/* <div className="flex flex-col mb-6">
+              <label className="text-sm2 text-modern-text mb-2">
+                Query Parameters (Optional)
+              </label>
+              <TextInput
+                value={queryParams}
+                onChange={(e) => setQueryParams(e)}
+                placeholder="e.g., id=1&name=John"
+                className="w-[250px]"
+              />
+            </div> */}
           </>
         ) : (
           <>
@@ -250,13 +342,11 @@ const DataBinding = () => {
             </div>
 
             {/* 쿼리 실행 버튼 */}
-            <button
-              className="bg-modern-point_10 text-modern-point p-2 rounded w-[250px] hover:bg-navy-btn_hover"
+            <LargeBtn
+              title={"Execute Query"}
               onClick={handleQueryExecute}
               disabled={!query.trim()}
-            >
-              Execute Query
-            </button>
+            />
           </>
         )}
       </div>
