+++ app/(monitoring)/d_/content/chartOption.tsx
