+++ app/(monitoring)/d_/content/rightSection.tsx
