+++ app/(monitoring)/d_/content/widgetOption.tsx
