+++ app/(monitoring)/dashboard_/content/mainSection.tsx
