+++ app/(monitoring)/dashboard_/page.tsx
@@ -1,0 +1,12 @@
+import { Suspense } from "react";
+import MainSection from "./content/mainSection";
+
+const DetailDashboard = () => {
+  return (
+    <Suspense>
+      <MainSection />
+    </Suspense>
+  );
+};
+
+export default DetailDashboard;
