+++ app/(monitoring)/detail/content/detailDashboard copy 2.tsx
