+++ app/(monitoring)/detail/content/detailDashboard copy 3.tsx
