+++ app/(monitoring)/detail/content/detailDashboard copy.tsx
