+++ app/(monitoring)/detail_/page.tsx
